ci: Fix clippy warnings and errors

+++ token-swap/program/src/curve/constant_price.rs
@@ -569,7 +569,7 @@ mod tests {
         ) {
             let curve = ConstantPriceCurve { token_b_price: token_b_price as u64 };
             let pool_token_amount = pool_token_amount as u128;
-            let pool_token_supply = pool_token_supply as u128;
+            let pool_token_supply = pool_token_supply;
             let swap_token_a_amount = swap_token_a_amount as u128;
             let swap_token_b_amount = swap_token_b_amount as u128;
             let token_b_price = token_b_price as u128;
